Split signature features into separate plugins
