+++ README.md
@@ -198,8 +198,13 @@ For inquiries: **doccropper@iltuoconsulenteit.it**
 
 Simple image or drawn signatures are available in all editions, but the Free edition keeps the watermark on exported PDFs.
 
-Signature functionality is packaged as a plugin under `plugins/signature` so it can evolve separately.  This folder now contains both the client script and the Python routes registered by `main.py`.
-> **Note** The mobile signing page includes a disclaimer that DocCropper and its authors accept no liability for illegal use. After scanning the QR code the phone fetches all pages so you can choose any page from the dropdown. Press **Finish** to lock the signatures. Contract-signing workflows may be provided as a Full edition plugin.
+Signature functionality is split into three optional plugins under `plugins/`:
+`sign` for local page stamping, `mobilesign` for signing from a smartphone and
+`remotesign` for Docuseal or other external services. The mobile signing page
+includes a disclaimer that DocCropper and its authors accept no liability for
+illegal use. After scanning the QR code the phone fetches all pages so you can
+choose any page from the dropdown. Press **Finish** to lock the signatures.
+Contract-signing workflows may be provided as a Full edition plugin.
 DocCropper can apply a personal signature in several ways:
 
 1. **Image Stamp** – Click the **Sign Page** button to open the signing panel,
