+++ plugins/mobilesign/README.md
@@ -1,0 +1,1 @@
+This folder contains the mobile signing plugin. It exposes a smartphone-friendly signing page via `/sign/<token>` and hooks into the web UI through `static/plugins/mobilesign.js`.
