+++ plugins/remotesign/README.md
@@ -1,0 +1,1 @@
+This plugin handles remote digital signing via Docuseal or a custom command.
