+++ plugins/sign/README.md
@@ -1,0 +1,1 @@
+This plugin provides the desktop signature tools. Signing is handled entirely in the browser so no API routes are needed.
