+++ plugins/sign/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ['register']
+
+def register(app, utils):
+    """Local signature plugin - no server routes required."""
+    pass
