--- plugins/signature/README.md
@@ -1,2 +1,0 @@
-This folder contains the optional signature plugin. It hooks into the web UI via
-`static/plugins/signature.js`. Future versions may ship the plugin separately.
