+++ static/index.html
@@ -187,6 +187,6 @@
     <script src="https://code.jquery.com/jquery-3.6.0.min.js" integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/sortablejs@1.15.2/Sortable.min.js"></script>
     <script type="module" src="/static/app.js"></script>
-    <script type="module" src="/static/plugins/signature.js"></script>
+    <script type="module" src="/static/plugins/mobilesign.js"></script>
 </body>
 </html>
